# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Pharma API",
    "description": "Week7 CSE 341 - Rune Larsen",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/meds/": {
      "get": {
        "tags": [
          "Meds"
        ],
        "description": "Endpoint to get all MEDS contacts",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Meds"
        ],
        "description": "Endpoint to create a Meds",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ClaimReference": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "412": {
            "description": "Precondition Failed"
          }
        }
      }
    },
    "/meds/{id}": {
      "get": {
        "tags": [
          "Meds"
        ],
        "description": "Endpoint to get one Meds",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "Meds"
        ],
        "description": "Endpoint to update all Meds",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "ClaimReference": {
                  "example": "any"
                },
                "Name": {
                  "example": "any"
                },
                "FillDate": {
                  "example": "any"
                },
                "Prescription": {
                  "example": "any"
                },
                "Rx": {
                  "example": "any"
                },
                "Qty": {
                  "example": "any"
                },
                "Prescriber": {
                  "example": "any"
                },
                "Pharmacist": {
                  "example": "any"
                },
                "NDC": {
                  "example": "any"
                },
                "Insurance": {
                  "example": "any"
                },
                "Claim": {
                  "example": "any"
                },
                "Price": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "412": {
            "description": "Precondition Failed"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "Meds"
        ],
        "description": "Endpoint to delete a Meds",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}